[Feature] support dump slow driver stats (#33806)
